got ssl stuff working
